refactor(Scroller): move Scroll utility to base package
