+++ packages/base/src/Scroller/useScrollElement.ts
@@ -1,5 +1,5 @@
 import { useContext, useEffect } from 'react';
-import { ScrollContext } from '@ui5/webcomponents-react/lib/ScrollContext';
+import { ScrollContext } from '@ui5/webcomponents-react-base/lib/ScrollContext';
 
 export const useScrollElement = (id, htmlRef, options = {}) => {
   const { registerElement, unregisterElement } = useContext(ScrollContext);
