+++ packages/base/src/interfaces/IScroller.ts
@@ -1,0 +1,7 @@
+import { Event } from '@ui5/webcomponents-react-base/lib/Event';
+
+export interface IScroller {
+  scroll: (e: Event) => void;
+  scrollToElementById: (id: string, scrollOffset?: number) => void;
+  scrollToTop: () => void;
+}
