+++ packages/base/src/lib/ScrollContext.ts
@@ -1,0 +1,2 @@
+import { ScrollContext } from '../Scroller/ScrollContext';
+export { ScrollContext };
