+++ packages/base/src/lib/ScrollLink.ts
@@ -1,0 +1,2 @@
+import { ScrollLink } from '../Scroller/ScrollLink';
+export { ScrollLink };
