+++ packages/base/src/lib/Scroller.ts
@@ -1,0 +1,2 @@
+import { Scroller } from '../Scroller/Scroller';
+export { Scroller };
