+++ packages/base/src/lib/useScrollElement.ts
@@ -1,0 +1,2 @@
+import { useScrollElement } from '../Scroller/useScrollElement';
+export { useScrollElement };
