+++ packages/main/src/components/ObjectPage/ObjectPageAnchorButton.tsx
@@ -8,7 +8,7 @@ import { ObjectPageMode } from '@ui5/webcomponents-react/lib/ObjectPageMode';
 import { PlacementType } from '@ui5/webcomponents-react/lib/PlacementType';
 import { Popover } from '@ui5/webcomponents-react/lib/Popover';
 import { StandardListItem } from '@ui5/webcomponents-react/lib/StandardListItem';
-import { ObjectPageLink } from './scroll/ObjectPageLink';
+import { ScrollLink } from '@ui5/webcomponents-react-base/lib/ScrollLink';
 
 interface ObjectPageAnchorPropTypes {
   section: any;
@@ -118,20 +118,20 @@ export const ObjectPageAnchorButton: FC<ObjectPageAnchorPropTypes> = (props) => 
     }
 
     return (
-      <ObjectPageLink
+      <ScrollLink
         key={item.props.id}
         id={`ObjectPageSubSection-${item.props.id}`}
         scrollOffset={collapsedHeader ? 45 : 0}
       >
         <StandardListItem data-key={item.props.id}>{item.props.title}</StandardListItem>
-      </ObjectPageLink>
+      </ScrollLink>
     );
   };
 
   let sectionSelector = null;
   if (mode === ObjectPageMode.Default) {
     sectionSelector = (
-      <ObjectPageLink
+      <ScrollLink
         id={`ObjectPageSection-${section.props.id}`}
         onSetActive={onScrollActive}
         activeClass={classes.selected}
@@ -139,7 +139,7 @@ export const ObjectPageAnchorButton: FC<ObjectPageAnchorPropTypes> = (props) => 
         scrollOffset={45}
       >
         <span className={classes.button}>{section.props.title}</span>
-      </ObjectPageLink>
+      </ScrollLink>
     );
   } else {
     sectionSelector = (
