--- packages/main/src/lib/ScrollContext.ts
@@ -1,2 +1,0 @@
-import { ScrollContext } from '../components/ObjectPage/scroll/ScrollContext';
-export { ScrollContext };
