+++ tsconfig.json
@@ -16,7 +16,7 @@
     "paths": {
       "@shared/*": ["shared/*"],
       "@ui5/webcomponents-react/lib/*": ["node_modules/@ui5/webcomponents-react/src/lib/*"],
-      "@ui5/webcomponents-react-base/lib/*": ["node_modules/@ui5/webcomponents-react-base/src/lib/*"],
+      "@ui5/webcomponents-react-base/*": ["node_modules/@ui5/webcomponents-react-base/src/*"],
       "@ui5/webcomponents-react-charts/lib/*": ["node_modules/@ui5/webcomponents-react-charts/src/lib/*"]
     },
     "moduleResolution": "node",
